datavault-gcp-pipeline/requirements.txt
Apache Beam==2.34.0
google-cloud-storage==2.5.0
google-cloud-bigquery==2.34.0
google-cloud-pubsub==2.10.0
pandas==1.3.5
numpy==1.21.6